refactor: extract CLI into standalone whitenoise-cli workspace crate (Phase 19)
Moves all CLI code out of the whitenoise core library and into a new
`crates/whitenoise-cli` workspace member. The core library no longer
carries CLI-specific dependencies (clap, dirs, rpassword) directly —
they live exclusively in the CLI crate.

Changes:
- New `crates/whitenoise-cli` crate with its own Cargo.toml
- Moves `src/cli/` → `crates/whitenoise-cli/src/cli/`
- Moves `src/bin/wn.rs` and `src/bin/wnd.rs` → `crates/whitenoise-cli/src/bin/`
- Moves `tests/cli_e2e.rs` and `tests/cli_relay_control_e2e.rs` → `crates/whitenoise-cli/tests/`
- Removes `dirs` and `rpassword` from core workspace deps
- Retains `cli = []` as an empty feature flag in core (used by
  `src/whitenoise/mod.rs` for macOS keyring store selection)
- Makes `Relay::defaults()` pub so the CLI crate can call it
- Adds `UserSearchParams` and `ChatListItem` to `lib.rs` re-exports
  so CLI code uses clean top-level paths instead of deep module paths
