+++ crates/whitenoise-cli/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "whitenoise-cli"
+version = "0.2.1"
+description = "Command-line interface for the Whitenoise daemon"
+authors = ["White Noise Authors"]
+edition = "2024"
+rust-version = "1.90.0"
+
+[dependencies]
+whitenoise = { path = "../..", features = ["cli"] }
+clap = { version = "4.5.37", features = ["derive"] }
+dirs = "6"
+rpassword = "7"
+tokio = { version = "1.44", features = ["full"] }
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
+thiserror = "2.0.12"
+mdk-core = { version = "0.7.1", git = "https://github.com/marmot-protocol/mdk", rev = "8a8d06c19d617394b0e268f5f9adfad34db3cdf6", features = [
+    "mip04",
+    "mip05",
+] }
+nostr-sdk = { version = "0.44", features = [
+    "nip04",
+    "nip44",
+    "nip47",
+    "nip59",
+] }
+chrono = { version = "0.4.40", features = ["serde"] }
+tracing = "0.1"
+tempfile = "3.19.1"
+hex = "0.4"
+
+[dev-dependencies]
+whitenoise = { path = "../..", features = ["cli", "integration-tests"] }
+tokio = { version = "1.44", features = ["test-util"] }
+
+[[bin]]
+name = "wn"
+path = "src/bin/wn.rs"
+
+[[bin]]
+name = "wnd"
+path = "src/bin/wnd.rs"
