+++ crates/whitenoise-cli/src/cli/output.rs
