+++ crates/whitenoise-cli/tests/support/cli.rs
