+++ docs/session-projection-implementation-plan.md
@@ -588,7 +588,7 @@ joins across account DB and shared DB.
 
 ### Phase 19: CLI crate extraction (~500 LOC of Cargo.toml + module moves)
 
-- [ ] Not started
+- [x] Completed in PR #XXX
 
 <details>
 <summary>Details</summary>
